LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-no-upgrade-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-no-upgrade-header' failed: Error during WebSocket handshake: 'Upgrade' header is missing
Test that WebSocket handshake fails if Upgrade header field isn't reflected by the server.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
